backend/requirements.txt
flask
google-cloud-storage
google-api-python-client
sqlalchemy
gunicorn
pymysql
